requirements.txt
flask
dnspython
publicsuffix2